refactor(contracts/deploy): add `PrimaryTokensTestnets` script
It deploys primary tokens on testnets.
Also changed `deploy.sh` script to don't call that script if the `TYPE`
is `main`.
